pyproject.toml
[project]
name = "torq-llm-engineer"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "asyncer==0.0.7",
    "attrs==25.3.0",
    "bidict==0.23.1",
    "certifi==2024.8.30",
    "chainlit==1.3.2",
    "charset-normalizer==3.4.0",
    "chevron==0.14.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.15",
    "distro==1.9.0",
    "fastapi==0.115.5",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "googleapis-common-protos==1.66.0",
    "greenlet==3.2.3",
    "grpcio==1.68.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.0",
    "huggingface-hub==0.31.1",
    "idna==3.10",
    "importlib-metadata==8.5.0",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "jiter==0.8.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "langchain==0.3.26",
    "langchain-core==0.3.66",
    "langchain-openai==0.3.24",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.8",
    "langgraph-checkpoint==2.1.0",
    "langgraph-prebuilt==0.2.2",
    "langgraph-sdk==0.1.70",
    "langsmith==0.4.1",
    "lazify==0.4.0",
    "literalai==0.0.623",
    "markupsafe==3.0.2",
    "marshmallow==3.23.1",
    "multidict==6.4.3",
    "mypy-extensions==1.0.0",
    "nest-asyncio==1.6.0",
    "numpy==1.26.4",
    "openai==1.86.0",
    "opentelemetry-api==1.28.2",
    "opentelemetry-exporter-otlp==1.28.2",
    "opentelemetry-exporter-otlp-proto-common==1.28.2",
    "opentelemetry-exporter-otlp-proto-grpc==1.28.2",
    "opentelemetry-exporter-otlp-proto-http==1.28.2",
    "opentelemetry-instrumentation==0.49b2",
    "opentelemetry-proto==1.28.2",
    "opentelemetry-sdk==1.28.2",
    "opentelemetry-semantic-conventions==0.49b2",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==23.2",
    "pluggy==1.6.0",
    "propcache==0.3.1",
    "protobuf==5.29.0",
    "pydantic==2.10.1",
    "pydantic-core==2.27.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pytest==8.4.0",
    "python-dotenv==1.0.1",
    "python-engineio==4.10.1",
    "python-multipart==0.0.9",
    "python-socketio==5.11.4",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.24.0",
    "simple-websocket==1.1.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "starlette==0.41.3",
    "syncer==2.0.3",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "tomli==2.2.1",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "uptrace==1.28.2",
    "urllib3==2.2.3",
    "uvicorn==0.25.0",
    "watchfiles==0.20.0",
    "wrapt==1.17.0",
    "wsproto==1.2.0",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]
